Import the cflags cmake definitions from aws-c-common (with some changes)
This also fixes some compiler warnings exposed by the new cflags.

+++ source/enc_context.c
@@ -92,7 +92,7 @@ int aws_cryptosdk_context_serialize(struct aws_allocator *alloc,
     output->len = length;
     struct aws_byte_cursor cur = aws_byte_cursor_from_buf(output);
 
-    if (!aws_byte_cursor_write_be16(&cur, num_elems)) goto WRITE_ERR;
+    if (!aws_byte_cursor_write_be16(&cur, (uint16_t)num_elems)) goto WRITE_ERR;
 
     for (size_t idx = 0; idx < num_elems; ++idx) {
         struct aws_hash_element elem;
@@ -102,9 +102,9 @@ int aws_cryptosdk_context_serialize(struct aws_allocator *alloc,
         }
         const struct aws_string * key = (const struct aws_string *)elem.key;
         const struct aws_string * value = (const struct aws_string *)elem.value;
-        if (!aws_byte_cursor_write_be16(&cur, key->len)) goto WRITE_ERR;
+        if (!aws_byte_cursor_write_be16(&cur, (uint16_t)key->len)) goto WRITE_ERR;
         if (!aws_byte_cursor_write_from_whole_string(&cur, key)) goto WRITE_ERR;
-        if (!aws_byte_cursor_write_be16(&cur, value->len)) goto WRITE_ERR;
+        if (!aws_byte_cursor_write_be16(&cur, (uint16_t)value->len)) goto WRITE_ERR;
         if (!aws_byte_cursor_write_from_whole_string(&cur, value)) goto WRITE_ERR;
     }
     aws_array_list_clean_up(&elems);
